Move tests for forPackage to correct folder.
